Moved OBC communication to separate library
