+++ libs/obc/communication/CMakeLists.txt
@@ -1,0 +1,16 @@
+set(NAME obc_communication)
+
+set(SOURCES
+    communication.cpp
+)
+
+add_library(${NAME} STATIC ${SOURCES})
+
+target_link_libraries(${NAME} 
+	base
+	comm
+	telecommand_handling
+)
+
+target_include_directories(${NAME} INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/Include)
+target_include_directories(${NAME} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/Include/obc)
